Compendium of changes before the 3/5 order
Some thoughts on what has changed:

1. Now using Image Boarder from Frank to create mirrors.
2. Now using Saturation from Frank to pump up the saturation.
3. Additional card sets added for now.

I want to start to improve these scripts to be more usable for a
broader set of the community, but now is not that time. Really,
this is a good stopping point of functionality as it has been
broadly expanded beyond the original scope of the branch.