April modifications
Tried a few different things. Despeckle on GIMP is an interesting one,
but causes some card text to be lost. Also decided that the saturation
effect was sufficient for brightening the card image (though we will
see with the resulting product.)

Removing the CMYK print pattern is the real challenge here.  Despeckle
seemed to do a good job, but made plenty of the text unreadable.  Just
doing a bunch of blur also seemed insufficient.

Lastly, I noticed that much of the bluriness on the cards was likely
caused by resizing on the factory side.  Here, an attempt is made to
reduce the bluring by doing the resize with our own algorithms.